Ginette can now delete old messages and speak with you !

+++ reactions.py
@@ -26,7 +26,8 @@ def get_trashcan_listener(bot):
         message = await MessageConverter().convert(ctx,f"{channel_id}-{message_id}")
         if payload.member.bot or message.author.id != ctx.bot.user.id:
             return
-        await message.delete()
+        if payload.emoji.name == Emoji.wastebasket:
+            await message.delete()
     return trashcan_listener
 
 
@@ -35,7 +36,7 @@ async def register_post_reactions(ctx, message, board, links):
     async def links_handler(reaction, user):
         if user.bot or reaction.message.id != message.id:
             return
-        reac_no = int(reaction.emoji[0])
+        reac_no = int(reaction.emoji[0]) if '0' <= reaction.emoji[0] <= '9' else -1
         if len(links) > reac_no > -1:
             await disbot.post(ctx, board, links[0], links[reac_no])
